tested api

+++ restaurant/models.py
@@ -20,7 +20,7 @@ class Booking(models.Model):
 class MenuItem(models.Model):
     title = models.CharField(max_length=255)
     price = models.DecimalField(max_digits=6, decimal_places=2)
-    inventory = models.SmallIntegerField()
+    inventory = models.IntegerField()
 
     def get_item(self):
         return f'{self.title} : {str(self.price)}'
